[ALPS04004493] mt6771 clk: fix check patch
[Detail] fix two warnings common code

MTK-Commit-Id: fdd59f7696abdd5498620cc31eb14a54a7f85bdd

CR-Id: ALPS04004493
Feature: Power Management
(cherry-pick from 58c05e7d49f59876238bc75faf9d395c151a03f3)

+++ drivers/clk/mediatek/clk-mtk.h
@@ -69,7 +69,7 @@ struct mtk_composite {
 	const char *name;
 	const char * const *parent_names;
 	const char *parent;
-	unsigned flags;
+	unsigned long flags;
 
 	uint32_t mux_reg;
 	uint32_t divider_reg;
